fix printf format warnings

+++ PLC_esp8266/main/WiFi/WiFiService_AccessPoint.cpp
@@ -133,7 +133,7 @@ void WiFiService::ap_connect_wifi_event_handler(void *arg,
 
     ESP_LOGI(TAG_WiFiService_AccessPoint,
              "connect client aid:%u, mac:" MACSTR ", mask:%s",
-             event->aid,
+             (unsigned int)event->aid,
              MAC2STR(event->mac),
              ap_event_arg->mac);
 
@@ -158,7 +158,7 @@ void WiFiService::ap_disconnect_wifi_event_handler(void *arg,
 
     ESP_LOGI(TAG_WiFiService_AccessPoint,
              "disconnect client aid:%u, mac:" MACSTR ", mask:%s",
-             event->aid,
+             (unsigned int)event->aid,
              MAC2STR(event->mac),
              ap_event_arg->mac);
 
@@ -178,15 +178,15 @@ void WiFiService::AddApClient(const char *ssid, t_mac mac) {
     }
     ESP_LOGI(TAG_WiFiService_AccessPoint,
              "AddApClient, ssid_cnt: %u, ssid_clients:%u",
-             (unsigned)ap_clients.size(),
-             (unsigned)it.first->second.size());
+             (unsigned int)ap_clients.size(),
+             (unsigned int)it.first->second.size());
 }
 
 size_t WiFiService::GetApClientsCount(const char *ssid) {
     std::lock_guard<std::mutex> lock(ap_clients_lock_mutex);
     auto it = ap_clients.find(ssid);
     bool found = it != ap_clients.end();
-    ESP_LOGD(TAG_WiFiService_AccessPoint, "FindApClient, found:%u", found);
+    ESP_LOGD(TAG_WiFiService_AccessPoint, "FindApClient, found:%u", (unsigned int)found);
     if (found) {
         return it->second.size();
     }
@@ -197,7 +197,7 @@ void WiFiService::RemoveApClient(const char *ssid, t_mac mac) {
     std::lock_guard<std::mutex> lock(ap_clients_lock_mutex);
     auto it = ap_clients.find(ssid);
     bool found = it != ap_clients.end();
-    ESP_LOGD(TAG_WiFiService_AccessPoint, "FindApClient, found:%u", found);
+    ESP_LOGD(TAG_WiFiService_AccessPoint, "FindApClient, found:%u", (unsigned int)found);
     if (found) {
         it->second.erase(mac);
         bool empty_ssid_to_be_delete = it->second.size() == 0;
@@ -205,11 +205,15 @@ void WiFiService::RemoveApClient(const char *ssid, t_mac mac) {
             ap_clients.erase(ssid);
         }
     }
-    ESP_LOGI(TAG_WiFiService_AccessPoint, "RemoveApClient, cnt:%u", (unsigned)ap_clients.size());
+    ESP_LOGI(TAG_WiFiService_AccessPoint,
+             "RemoveApClient, cnt:%u",
+             (unsigned int)ap_clients.size());
 }
 
 void WiFiService::RemoveApClients(const char *ssid) {
     std::lock_guard<std::mutex> lock(ap_clients_lock_mutex);
     ap_clients.erase(ssid);
-    ESP_LOGI(TAG_WiFiService_AccessPoint, "RemoveApClients, cnt:%u", (unsigned)ap_clients.size());
+    ESP_LOGI(TAG_WiFiService_AccessPoint,
+             "RemoveApClients, cnt:%u",
+             (unsigned int)ap_clients.size());
 }
